feat: finished curcial steps moving forward
The project has now multiple subjects that can be indexed into the
platform. That means we can separate content inside the project and
show it selected for the users to consume.

The project now has production bootstrap, popper, jquery and
fontawesome - all served through npm modules. That might simplify new
features to be added and the existing structure to be modified.

Many project patterns need to take form yet, but the structure is
already solid and secure to develop.

Signed-off-by: José Tobias <tobias.jc07@gmail.com>